# railway-backend/openapi.yaml
openapi: 3.0.3
info:
  title: All For Pets — Municipal Pet Registry API
  version: 1.0.0
  description: |
    REST API powering the AFP web + Expo mobile app.

    ### Authentication
    - `POST /api/auth/login` returns `token` (JWT access, 15 min) and
      `refreshToken` (opaque, device-bound, 1–30 days).
    - Send `Authorization: Bearer <token>` on every protected call.
    - On 401 → call `POST /api/auth/refresh` with the refresh token; a **new**
      refresh token is returned (rotation) and the old one is revoked.
    - Mobile clients should send `X-Device-Label` and `X-Expo-Push-Token`
      headers on login/register/refresh so we can list / revoke sessions and
      push notifications.

    ### Caching
    - Public `/api/geo/*` endpoints return `ETag` + `Cache-Control`.
      Send `If-None-Match` to get a 304 with an empty body.

    ### File uploads (mobile)
    - Prefer the `/upload-photo/sign` + `/upload-photo/confirm` flow so the
      binary goes directly to S3/Azure/Cloudinary without touching the API.
servers:
  - url: https://afp.up.railway.app
    description: Production (Railway)
  - url: http://localhost:3000
    description: Local development (Node)
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      security: []
      summary: Register a citizen account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, mobile, password]
              properties:
                name:     { type: string, example: "Tapan Singh" }
                mobile:   { type: string, example: "9810000000" }
                email:    { type: string, format: email, nullable: true }
                password: { type: string, format: password, minLength: 8 }
                address:  { type: string, nullable: true }
                cityId:   { type: integer, nullable: true }
                nigamId:  { type: integer, nullable: true }
                zoneId:   { type: integer, nullable: true }
                wardId:   { type: integer, nullable: true }
      responses:
        "201": { $ref: "#/components/responses/AuthOk" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Login with mobile or email + password
      parameters:
        - $ref: "#/components/parameters/DeviceLabel"
        - $ref: "#/components/parameters/ExpoPushToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, password]
              properties:
                identifier: { type: string, description: "Mobile or email" }
                password:   { type: string, format: password }
                remember:   { type: boolean, description: "30-day refresh vs 1-day" }
      responses:
        "200": { $ref: "#/components/responses/AuthOk" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: Inactive account or geo scope disabled }
  /api/auth/refresh:
    post:
      tags: [Auth]
      security: []
      summary: Exchange a refresh token for a fresh access token (rotates the refresh token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200": { $ref: "#/components/responses/AuthOk" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Revoke a single device or all sessions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string, description: "Revoke this device only" }
                allDevices:   { type: boolean, description: "Revoke every session (needs Bearer)" }
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { type: object, properties: { revoked: { type: integer } } } } }
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Current user
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { $ref: "#/components/schemas/User" } } }
    put:
      tags: [Auth]
      summary: Update own profile / change password
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:            { type: string }
                mobile:          { type: string }
                email:           { type: string, format: email }
                address:         { type: string }
                currentPassword: { type: string, format: password }
                newPassword:     { type: string, format: password }
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { $ref: "#/components/schemas/User" } } }
...
  /api/auth/sessions/{id}:
    delete:
      tags: [Auth]
      summary: Revoke a specific session (device)
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { type: object, properties: { revoked: { type: integer } } } } }

  # ───────────── Geo ─────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    DeviceLabel:
      in: header
      name: X-Device-Label
      required: false
      schema: { type: string, example: "iPhone 15 · Tapan" }
      description: Displayed in the "Manage devices" screen after login.
    ExpoPushToken:
      in: header
      name: X-Expo-Push-Token
      required: false
      schema: { type: string }
      description: Stored per session so push notifications can be sent to this device.
  responses:
    AuthOk:
      description: OK
      content:
        application/json:
          schema:
            type: object
            required: [token, refreshToken, user]
            properties:
              token:        { type: string, description: "JWT access token, 15 min" }
              refreshToken: { type: string, description: "Opaque refresh token" }
              user:         { $ref: "#/components/schemas/User" }
    Unauthorized:
      description: Invalid credentials or token
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    BadRequest:
      description: Bad input
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Conflict:
      description: Duplicate resource
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  schemas:
    Error:
      type: object
      properties:
        error:   { type: string }
        details: { type: string, nullable: true }
    User:
      type: object
      properties:
        id:          { type: integer }
        name:        { type: string }
        mobile:      { type: string }
        email:       { type: string, nullable: true }
        address:     { type: string, nullable: true }
        role:        { type: string, enum: [citizen, ward_admin, zone_admin, nigam_admin, city_admin, super_admin] }
        city_id:     { type: integer, nullable: true }
        city_name:   { type: string, nullable: true }
        nigam_id:    { type: integer, nullable: true }
        nigam_name:  { type: string, nullable: true }
        zone_id:     { type: integer, nullable: true }
        zone_name:   { type: string, nullable: true }
        ward_id:     { type: integer, nullable: true }
        ward_number: { type: string, nullable: true }
        is_active:   { type: boolean }
        created_at:  { type: string, format: date-time }
...